pyproject.toml
[tool.poetry]
name = "lin-eval-commander"
version = "0.0.1"
description = ""
authors = ["Daniel Trnka <daniel.trnka@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.8"
plin-linux = "^0.0.3"
ldfparser = "^0.25.0"
cmd2 = "^2.5.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
